Classifier/requirements.txt
我的库版本，其他版本应该也可以

numpy=1.17.3
py-opencv=3.4.2
pytorch = 1.3.1